requirements.txt
apache-beam
pytest
python-dateutil~=2.8.2
setuptools~=65.5.0